Add #undefs for posix source symbols when including papi.h
Validation fails when validating with PAPI support (i.e. GhcRtsWithPapi  = YES
in validate.mk).  The problem is that the posix symbols are defined by a header
included from papi.h. Compilation then fails because these symbols are
redefined in PosixSource.h. 

This patch adds an undefine for the posix symbols after including papi.h and
before including PosixSource.h. The #undefines are localized to Papi.c since
that is the only case where they are getting defined twice.