Refactor the command-line argument parsing (again)
This change allows the client of CmdLineParser a bit more flexibility,
by giving him an arbitrary computation (not just a deprecation
message) for each flag.  

There are several clients, so there are lots of boilerplate changes.

Immediate motivation: if RTS is not profiled, we want to make 
Template Haskell illegal.  That wasn't with the old setup.