Remove GHC.extendGlobalRdrScope, GHC.extendGlobalTypeScope
These functions were added by

   Tue Apr 18 03:36:06 BST 2006  Lemmih <lemmih@gmail.com>
   * Make the initial rdr and type scope available in the ghc-api

The are extremely dubious, because they extend the Rdr and Type
env for every compilation.  The right thing to do is to use
the InteractiveContext for temporary extensions.

So far as we know, no one uses them.  And if they are being used
it's probably a mistake.  So we're backing them out.