Fix initialisation of strictness in the demand analyser
Previously, the demand analyser assumed that every binder 
starts off with no strictness info.  But now that we are
preserving strictness on nesting bindings in interface files,
that assumption is no longer correct, because an inlined function
might have a nested binding with strictness set.

So we need to know when we are in the initial sweep, so that we can
set the strictness to 'bottom'. 

See Note [Initialising strictness]