Some infrastruture for lambda-lifting
This stuff should have no effect but it sets things
up so that we can try floating out lambdas of n value
arguments.

The new (secret) flag is -ffloatt-lam-args=n.

This is *not* working yet, but it's got tangled up with
other stuff I want to commit, and it does no harm.