Fix #4829 (build does not respect --with-gcc option)
This patch fixes what seems to be the last problem with the --with-gcc
option.  On OS X, we need to pass the path to gcc to dtrace as the
preprocessor.  (Internally, dtrace on OS X sets the default preprocessor
to /usr/bin/gcc.)  ATM, dtrace is only supported on OS X, so we don't
need any conditionalization.  If dtrace is ported to other platforms,
we might need to change this. However, usage on other platforms will
probably be similar to OS X, since many of Apple's changes are to
use the gnu toolchain instead of the Sun toolchain.