Make Template Haskell classInstances function return [ClassInstance]
This is a recently-introduce function, which was returning
a [Name], being the names of the dfuns.  But what you really
want (obviously!) is the ClassInstances, and we have a TH type
for that.

This is an API change, so don't merge into GHC 7.0.  But it's
a new part of TH which is still settling down.

Fixes Trac #4863.