Add .gitignore.
This is already on the main branch but I chose the base for my branch
at a point before it was created.  I did this, because this  version of
the repo (roughly) matches the version which was in darcs, when I last
merged my changes so, hopefully, I can import my darcs patches without
problems.