Treat type operators as constants, not variables.
Modifies the parser to treat all infix operators at the type level
as constants, not type variables.

This allows us to make definitions such as this:

>  data a + b = Test

Change to import/exports
~~~~~~~~~~~~~~~~~~~~~~~~

With this change, writing (+) in an import/export specification is
ambiguous because it is not clear if we should be importing/exporting
the value level (+) or the type level one, or both.

We solve this as follows:

- Specifications which have subrodinates are aways types.
  For example, T(), F(..), and (+)() all refer to types.

- Specification which do not have subbordinates remain unchanged:
  - variable names refer to values:   f, (+)
  - constructor names refer to types: T, F, (:+)

- Because (+)() looks odd, we also add a bit of new syntax.  Writing:

> type (+)

is an abbreviation for:

> (+)()

(i.e., it imports/exports a type or class without any of its subordinates).

To avoid clutter, we also allow multiple type/class constructors to be
mentioned in a single 'type' specificatoin.  For example:

>  type (+) (-) TypeNat

imports/exports all of the specified types/classes.