Add basic support for number type literals.
We add a new kind, Nat, inhabited by a family of types,
one for each natural number:

0, 1, 2 .. :: Nat

In terms of GHC's sub-kind relation, Nat is only a sub-kind of itself.

The numeric types are empty because there are no primitives of these
types, and the kind "Nat" is not related to *, the kind of types which
can be defined in Haskell programs.