Fix #4867 (updated; corrects address calculation)
This is a corrected fix for ticket #4867, "ghci displays negative floats
incorrectly".  The previous patch sometimes gave incorrect offset to values
in the __const section of the __TEXT segment.  The new patch arranges a zero
fixup for non-external, not-global offset table signed relocations.  This
is apparently what is required, though documentation on this point is scarce.

With this change Doubles are negated properly, because the sign bit mask
is loaded from the correct offset.  This was tested both on HEAD and the 7.0
branch.