Prefer builtin commands to macros in GHCi command resolution (#3858)
Current precedence rules in GHCi are that:

 * User macros are *always* preferred in command resolution
 * User macros that are defined earlier are preferred to those that
   are defined earlier on
 * Builtin commands have lowest precedence

However this caused user confusion because e.g. defining a macro beginning
with "i" would override the standard :info command whenever the user typed
the abbreviated command :i.

The new precedence rules are based on the view that things defined earlier
are always preferred to things defined later. The builtin commands are logically
defined earliest of all (when GHCi starts) so they always take precedence.